services/receptionist/requirements.txt
pipecat-ai[daily,whisper,silero,anthropic,piper]==1.0.0
pipecat-ai-flows>=0.0.14
anthropic>=0.40.0
pyyaml>=6.0
python-dotenv>=1.0
fastapi>=0.110
uvicorn[standard]>=0.29
